requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==37.0.4
defusedxml==0.7.1
deprecated==1.2.13
diff-match-patch==20200713; python_version >= '2.7'
django-allauth==0.51.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-environ==0.9.0
django-filter==22.1
django-import-export==2.9.0
django-microsoft-auth==2.4.1
django-oauth-toolkit==2.1.0
django-rest-auth==0.9.5
django==3.2
djangorestframework-jwt==1.10.0
djangorestframework-simplejwt==5.2.0
djangorestframework==3.11.0
drf-social-oauth2==1.2.1
drf-yasg==1.21.3
ecdsa==0.18.0
et-xmlfile==1.1.0; python_version >= '3.6'
Faker==15.3.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.2
jwcrypto==1.3.1
markdown==3.4.1
markuppy==1.14
markupsafe==2.1.1
mock==4.0.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3; python_version >= '3.6'
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pyjwt==1.7.1
pyparsing==3.0.9
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
pyyaml==6.0
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.2
tablib[html,ods,xls,xlsx,yaml]==3.2.1; python_version >= '3.7'
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1